### 2025-02-20 #7822 @editorialbot
[PRE REVIEW]: Scanbuddy: fMRI motion plotting and SNR estimation at scan acquisition
**Submitting author:** @danielasay (<a href="http://orcid.org/0000-0002-6691-7706">Daniel Asay</a>)
**Repository:** https://github.com/harvard-nrg/scanbuddy
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.3
