### 2025-03-04 d1ef101
fix link to docs in readme
