### 2025-03-17 #7822 @Kevin-Mattheus-Moerman
@danielasay - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
