### 2025-04-07 bd04dcb
fix paper typos
