+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: citations.bib
 
 Functional magnetic resonance imaging (fMRI) is a powerful research and clinical tool for in vivo imaging of human brain function. `Scanbuddy` is containerized software that produces motion plots at the time of scan acquisition. Users should have a machine separate from the scanner host PC with its own display monitor that is capable of running Linux and Docker. Users can set up auto exporting from the scanner host PC via a Samba share mount for BOLD scans. With auto exporting, the scanner will automatically send reconstructed DICOMs to both the scanner host PC and the `Scanbuddy` machine. `Scanbuddy` uses the excellent dcm2niix [@li_first_2016] and AFNI [@cox_afni:_1996] packages for DICOM conversion and computing motion estimates. `Scanbuddy` was developed in a linux environment and can handle Repetition Times as fast as 0.6 seconds. 
 
-`Scanbuddy` also provides an estimate of the Signal-to-Noise Ratio (SNR) with the motion plots to estimate data quality. Scanbuddy does not save motion plots by default and does not persist data on its host machine. `Scanbuddy` will create a new motion plot and compute a new SNR metric for every fMRI scan acquired. `Scanbuddy` supports multi-echo BOLD scans with the assumption that the second echo time (TE2) is the TE of interest. `Scanbuddy` is containerized with Docker and is available on GitHub Container Registry. `Scanbuddy` also checks head-coil elements and notifies users if a head-coil is missing expected elements (e.g. the head-coil is not plugged in properly). In many ways, Scanbuddy is the real-time acquisition version of the Automated Functional MRI Quality Assesment [@gabriele_automated_2014] software package with several additional features.
+`Scanbuddy` also provides an estimate of the Signal-to-Noise Ratio (SNR) with the motion plots to estimate data quality. `Scanbuddy` does not save motion plots by default and does not persist data on its host machine. `Scanbuddy` will create a new motion plot and compute a new SNR metric for every fMRI scan acquired. `Scanbuddy` supports multi-echo BOLD scans with the assumption that the second echo time (TE2) is the TE of interest. `Scanbuddy` is containerized with Docker and is available on GitHub Container Registry. `Scanbuddy` also checks head-coil elements and notifies users if a head-coil is missing expected elements (e.g. the head-coil is not plugged in properly). In many ways, `Scanbuddy` is the real-time acquisition version of the Automated Functional MRI Quality Assessment [@gabriele_automated_2014] software package with several additional features.
 
 
 # Statement of need
