### 2025-04-07 #7822 @sappelhoff
Thank you @danielasay, I am happy to take this on as an editor.

Do you have suggestions on who may be a suitable peer-reviewer? (taking into account our [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy))
