### 2025-04-11 #7822 @PeerHerholz
Hey @sappelhoff and everyone,

sorry for the late reply.

This sounds very interesting, but unfortunately, I'm rather busy for the next couple of weeks.
Thus, I couldn't do the review in a reasonable amount of time. Sorry!

Best, Peer
