### 2025-04-12 #7822 @sappelhoff
Thanks for the notes, Peer and Haykel. 🙏

👋 @paquiteau @cecilyen @jonbmartin would you be available and willing to review this submission for JOSS?
