### 2025-04-18 c258b58
Update example-config.yaml
