### 2025-04-21 #7822 @sappelhoff
👋 @bwheelz36, @tomelse

would you be available and willing to review this for JOSS?

---

Also: @nmuncy @cecilyen @jonbmartin, a kind reminder in case this has fallen through in your notifications. We would still like to hear from you in case you'd be willing to peer review this submission!
