### 2025-04-28 150e907
add feature that checks for table position with 32 channel head coils
