+++ scanbuddy/proc/params.py
@@ -8,7 +8,8 @@ class Params:
     def __init__(self, config, broker=None):
         self._config = config.find_one('$.params', dict())
         self._broker = broker
-        self._checked = False
+        self._coil_checked = False
+        self._table_checked = False
         pub.subscribe(self.listener, 'params')
         pub.subscribe(self.reset, 'reset')
 
@@ -16,8 +17,11 @@ class Params:
         self._checked = False
 
     def listener(self, ds):
-        if self._checked:
-            logger.info(f'already checked an instance from series {ds.SeriesNumber}')
+        if self._coil_checked:
+            logger.info(f'already checked coil from series {ds.SeriesNumber}')
+            return
+        if self._table_checked:
+            logger.info(f'already checked table table position from {ds.SeriesNumber}')
             return
         for item in self._config:
             args = self._config[item]
@@ -25,7 +29,7 @@ class Params:
             f(ds, args)
 
     def coil_elements(self, ds, args):
-        self._checked = True
+        self._coil_checked = True
         patient_name = ds.get('PatientName', 'UNKNOWN PATIENT')
         series_number = ds.get('SeriesNumber', 'UNKNOWN SERIES')
         receive_coil = self.findcoil(ds)
@@ -45,6 +49,34 @@ class Params:
                 logger.info(f'publishing message to message broker')
                 self._broker.publish('scanbuddy_messages', message)
                 break
+
+    def table_position(self, ds, args):
+        self._table_checked = True
+        patient_name = ds.get('PatientName', 'UNKNOWN PATIENT')
+        series_number = ds.get('SeriesNumber', 'UNKNOWN SERIES')
+        table_position = self.find_table_position(ds)
+        receive_coil = self.findcoil(ds)
+        message = args['message'].format(
+            SESSION=patient_name,
+            SERIES=series_number,
+            TABLE_POSITION=table_position,
+        )
+        for bad in args['bad']:
+            a = ( receive_coil, table_position )
+            b = ( bad['receive_coil'], bad['table_position'] )
+            logger.info(f'checking if {a} == {b}')
+            if a == b:
+                logger.warning(message)
+                logger.info('publishing message to message broker')
+                self._broker.publish('scanbuddy_messages', message)
+                break
+
+
+    def find_table_position(self, ds):
+        seq = ds[(0x5200, 0x9230)][0]
+        seq = seq[(0x0021, 0x11fe)][0]
+        return seq[(0x0021, 0x1145)].value[-1]
+
 
     def findcoil(self, ds):
         seq = ds[(0x5200, 0x9229)][0]
