### 2025-04-29 d9a2550
pin specific mamba version in dockerfile
