+++ Dockerfile
@@ -27,7 +27,7 @@ ENV PATH="${AFNI_PREFIX}:${PATH}"
 ARG D2N_PREFIX="/sw/apps/dcm2niix"
 ARG D2N_VERSION="1.0.20220720"
 ARG D2N_URI="https://github.com/rordenlab/dcm2niix/archive/refs/tags/v${D2N_VERSION}.zip"
-WORKDIR "/tmp"
+WORKDIR /tmp
 RUN dnf install -y unzip cmake gcc-c++ && \
     dnf --enablerepo=powertools install -y libstdc++-static && \
     curl -sL "${D2N_URI}" -o "dcm2niix_src.zip" && \
@@ -44,16 +44,18 @@ ENV PATH="${D2N_PREFIX}:${PATH}"
 
 # install miniforge
 ARG MFG_PREFIX="/sw/miniforge"
-WORKDIR "/tmp"
-RUN curl -sL "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh" -o "miniforge.sh"
-RUN bash "miniforge.sh" -b -p "${MFG_PREFIX}"
-RUN rm "miniforge.sh"
+WORKDIR /tmp
+RUN curl -sL "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-Linux-$(uname -m).sh" -o "miniforge.sh"
+RUN bash "miniforge.sh" -b -p "${MFG_PREFIX}" && \
+    rm "miniforge.sh"
 ENV PATH="${MFG_PREFIX}/bin:${PATH}"
 
-# install scanbuddy
+# install specific mamba version (1.5.12)
+RUN conda install -y -n base -c conda-forge mamba=1.5.12
+
+# install scanbuddy with python-freethreading
 RUN mamba create -y -n python3.13t --override-channels -c conda-forge python-freethreading
-RUN mamba env config vars set PYTHON_GIL=0 -n python3.13t
-ARG SB_VERSION="main"
+ARG SB_VERSION="0.2.4"
 RUN mamba run -n python3.13t --no-capture-output python3 -m pip install "git+https://github.com/harvard-nrg/scanbuddy.git@${SB_VERSION}"
 
 ENTRYPOINT ["mamba", "run", "-n", "python3.13t", "--no-capture-output", "start.py"]
