### 2025-04-29 cb951cc
add python_gil=0 environ
