### 2025-04-30 42efe24
pin version 3.13.1 of python inside docker image, apparent memory leak in version 3.13.3
