+++ scanbuddy/broker/redis.py
@@ -6,9 +6,10 @@ import redis.exceptions
 logger = logging.getLogger(__name__)
 
 class MessageBroker:
-    def __init__(self, host='localhost', port=6379):
-        self._host = host
-        self._port = port
+    def __init__(self, config, host='localhost', port=6379):
+        self._config = config
+        self._host = self._config.find_one('$.broker.host', default=host)
+        self._port = self._config.find_one('$.broker.port', default=port)
         self._conn = None
         self._uri = f'redis://{self._host}:{self._port}'
         self.connect()
