### 2025-05-06 11c012d
add additional logging and pin python version
