### 2025-05-06 29085cd
minor refactoring for docker compose
