### 2025-05-06 5d0c5dc
use local repo to build docker image
