### 2025-05-06 2a5414d
enable docker compose for development
