### 2025-05-06 51b0a32
add scanbuddy docker compose and some retry logic for dcm2niix
