+++ scripts/start.py
@@ -3,20 +3,23 @@
 import sys
 import time
 import logging
+import warnings
 from pubsub import pub
 from pathlib import Path
+from scanbuddy.proc.snr import SNR
 from argparse import ArgumentParser
-from scanbuddy.watcher.directory import DirectoryWatcher
-from scanbuddy.common import print_platform_info
+from scanbuddy.config import Config
 from scanbuddy.proc import Processor
+from scanbuddy.view.dash import View
 from scanbuddy.proc.volreg import VolReg
 from scanbuddy.proc.params import Params
-from scanbuddy.proc.snr import SNR
-from scanbuddy.view.dash import View
 from scanbuddy.broker.redis import MessageBroker
-from scanbuddy.config import Config
+from scanbuddy.common import print_platform_info
+from scanbuddy.watcher.directory import DirectoryWatcher
 
 logger = logging.getLogger('main')
+logging.captureWarnings(True)
+warnings.filterwarnings("ignore", category=DeprecationWarning)
 logging.basicConfig(level=logging.INFO)
 
 def main():
