### 2025-05-06 4a231bb
wait to check coil elements and table position until after 5th volume
