### 2025-05-07 7c48421
fix infinite loop when a new directory is created with the same name
