### 2025-05-07 a024660
allow setting debug display thought CLI or config file
