### 2025-05-07 2885f5a
add example HeadNeck_64 head coil error conditions, table positioning errors do not depend on the head coil
