+++ example-config.yaml
@@ -8,6 +8,7 @@ app:
         user: scanbuddy
         pass:
             env: SCANBUDDY_PASS
+    debug_display: false
 params:
     coil_elements:
         bad:
@@ -15,6 +16,10 @@ params:
               coil_elements:  HEA
             - receive_coil: Head_32
               coil_elements: HEP
+            - receive_coil: HeadNeck_64
+              coil_elements: HC2,4,6,7,NC2
+            - receive_coil: HeadNeck_64
+              coil_elements: HC2,4,6,7
         message: |
             Session: {SESSION}
             Series: {SERIES}
@@ -30,8 +35,7 @@ params:
 
     table_position:
         bad:
-            - receive_coil: Head_32
-              table_position  : -1250
+            - table_position: -1250
         message: |
             Session: {SESSION}
             Series: {SERIES}
