### 2025-05-07 25c0a94
table position is dependent on receive coil after all
