### 2025-05-07 8772305
add debug display to docker compose developer
