### 2025-05-07 5fc5d1b
add healthcheck to docker compose
