+++ example-docker-compose.yml
@@ -1,46 +1,0 @@
-## Example docker-compose.yml file
-
-# You can run a complete standalone system which serves an
-# SMB share for the scanner to write to. Then, the entire 
-# system is just a `docker compose up` away:
-
-services:
-  scanbuddy:
-    image: scanbuddy:v0.1.9
-    restart: unless-stopped
-    network_mode: host
-    depends_on:
-      - samba
-      - redis
-    environment:
-      # you might want passwords in a .env file
-      SCANBUDDY_PASS: some_password
-      SCANBUDDY_SESSION_KEY: some_password
-    volumes:
-      - /home/qc/scanbuddy:/scanbuddy
-      - /home/qc/scanbuddy-data:/data
-    command: ["-c", "/scanbuddy/config.yaml", "--folder", "/data", "--host", "0.0.0.0", "--port", "8080"]
-
-  redis:
-    image: redis:latest
-    restart: unless-stopped
-    network_mode: host
-
-  samba:
-    image: ghcr.io/servercontainers/samba
-    restart: unless-stopped
-    # you {c,sh}ould expose individual ports instead
-    network_mode: host
-    environment:
-      SAMBA_CONF_LOG_LEVEL: 3
-      WSDD2_DISABLE: 1
-      AVAHI_DISABLE: 1
-      NETBIOS_DISABLE: 1
-      GROUP_fakegroup: 20
-      ACCOUNT_scanbuddy: some_password
-      UID_scanbuddy: 501
-      GROUPS_scanbuddy: fakegroup
-      SAMBA_VOLUME_CONFIG_scanbuddy: "[scanbuddy]; path=/data; valid users = scanbuddy; guest ok = no; read only = no; browseable = yes"
-    volumes:
-      - /home/qc/scanbuddy-data:/data
-
