### 2025-05-07 fbdad8a
add make_arrays_zero method to free up memory after scan
