### 2025-05-07 7ab80e4
add logging to make_arrays_zero
