### 2025-05-07 f03552b
more docker compose goodness
