+++ scanbuddy-docker-compose/README.md
@@ -4,11 +4,10 @@ Customize `default.env`, save to a file named `.env`, then run
 docker compose up
 ```
 
-By default, this will start Scan Buddy, Redis, and NGINX reverse
-proxy.
+By default, this will start Scan Buddy, Redis, and Nginx.
 
 ## Samba
-If you want to start a Samba service container, add the `samba` profile
+To start a Samba service container, add the `samba` profile
 
 ```bash
 docker compose --profile samba up
@@ -16,7 +15,8 @@ docker compose --profile samba up
 
 ## Developers
 For developers, you may find it useful to mount your local development 
-copy of the Scan Buddy into the container at the appropriate location
+copy of Scan Buddy into the container at the appropriate location, rather 
+than having to rebuild the container
 
 ```bash
 docker compose -f docker-compose.yaml -f docker-compose-devel.yaml up
