+++ scanbuddy-docker-compose/docker-compose.yaml
@@ -17,7 +17,7 @@ services:
     memswap_limit: ${MEM_SWAP_LIMIT}
     command: ["-c", "/etc/scanbuddy.yaml", "--folder", "/data", "--host", "0.0.0.0", "--broker-host", "scanbuddy-broker"]
     healthcheck:
-      test: curl -sL -o /dev/null -u "scanbuddy:${SCANBUDDY_PASS}" --fail localhost:8080 
+      test: curl -sL -o /dev/null -u "${SCANBUDDY_USER}:${SCANBUDDY_PASS}" --fail localhost:8080
       interval: 30s
       timeout: 10s
       retries: 3
@@ -28,7 +28,7 @@ services:
     expose:
       - "6379"
     healthcheck:
-      test: ["CMD", "redis-cli", "ping"]
+      test: redis-cli ping
 
   scanbuddy-nginx:
     image: nginx:${NGINX_VERSION}
@@ -37,7 +37,7 @@ services:
       - ./nginx/nginx.conf:/etc/nginx/nginx.conf
       - ./nginx/logs:/var/log/nginx
     ports:
-      - "80:80"
+      - "${BIND_ADDR}:${BIND_PORT}:80"
     expose:
       - "80"
     environment:
@@ -45,7 +45,7 @@ services:
     depends_on:
       - scanbuddy-web
     healthcheck:
-      test: curl -sL -o /dev/null -u "scanbuddy:${SCANBUDDY_PASS}" --fail localhost
+      test: curl -sL -o /dev/null -u "${SCANBUDDY_USER}:${SCANBUDDY_PASS}" --fail localhost
       interval: 30s
       timeout: 10s
       retries: 3
