### 2025-05-08 728ab7d
print system resources when starting app, add cpu limit support to docker compose
