### 2025-05-08 032f0f9
change motion plot title. add logging and bug fix for deallocating snr arrays
