### 2025-05-08 f1bfbbf
add htop to image installation
