### 2025-05-09 54527cf
start watchers on all modalities inside of config file
