### 2025-05-09 68e169b
add scanbuddy version to initial table
