+++ scanbuddy/common/__init__.py
@@ -2,6 +2,7 @@ import os
 import sys
 import psutil
 import platform
+import scanbuddy
 from tabulate import tabulate
 
 def print_platform_info():
@@ -13,7 +14,8 @@ def print_platform_info():
         ['CPU Count', os.cpu_count()],
         ['System RAM', f'{rss:.2f} GB'],
         ['System Swap', f'{swap:.2f} GB'],
-        ['Python version', platform.python_version()],
+        ['Python Version', platform.python_version()],
+        ['Scanbuddy Version', scanbuddy.version()],
         ['GIL enabled', sys._is_gil_enabled()]
     ]
     print(tabulate(table, tablefmt='simple_grid'))
