### 2025-05-12 bbc715b
update version and add scanbuddy version to table printout
