+++ scanbuddy/common/__init__.py
@@ -13,7 +13,8 @@ def print_platform_info():
         ['CPU Count', os.cpu_count()],
         ['System RAM', f'{rss:.2f} GB'],
         ['System Swap', f'{swap:.2f} GB'],
-        ['Python version', platform.python_version()],
+        ['Python Version', platform.python_version()],
+	['Scanbuddy Version', scanbuddy.version()],
         ['GIL enabled', sys._is_gil_enabled()]
     ]
     print(tabulate(table, tablefmt='simple_grid'))
