### 2025-05-12 e81cb7b
docker compose: add scanbuddy-nginx to no_proxy environment variable, change default container version to 0.2.6
