### 2025-05-12 92f21a4
separate out the bold and localizer processors from the parent processor in __init__.py, create two directory watchers, one on the bold samba export and one on the localizer
