+++ scanbuddy/proc/bold.py
@@ -1,0 +1,449 @@
+import os
+import sys
+import pdb
+import time
+import math
+import json
+import shutil
+import psutil
+import logging
+import datetime
+import threading
+import numpy as np
+from pubsub import pub
+from pathlib import Path
+from sortedcontainers import SortedDict
+from scanbuddy.proc.snr import SNR
+
+logger = logging.getLogger(__name__)
+
+class BoldProcessor:
+    def __init__(self, config, debug_display=False):
+        self.reset()
+        self._config = config
+        self._debug_display = self._config.find_one('$.app.debug_display', default=debug_display)
+        pub.subscribe(self.reset, 'reset')
+        pub.subscribe(self.listener, 'bold-proc')
+
+    def reset(self):
+        self._instances = SortedDict()
+        self._slice_means = SortedDict()
+        self.make_arrays_zero('reset')
+        pub.sendMessage('plot_snr', snr_metric=str(0.0))
+        logger.debug('received message to reset')
+
+    def getsize(self, obj):
+        size_in_bytes = sys.getsizeof(obj)
+        return size_in_bytes
+
+    def get_size_slice_means(self):
+        total_size = 0
+        for key in self._slice_means:
+            slice_means = self._slice_means[key]['slice_means']
+            total_size += slice_means.nbytes
+        return total_size
+
+    def get_size_mask(self):
+        total_size = 0
+        for key in self._slice_means:
+            mask = self._slice_means[key]['mask']
+            if mask is not None:
+                mb = mask.nbytes / (1024**2)
+                shape = mask.shape
+                logger.info(f'mask for instance {key} is dtype={mask.dtype}, shape={shape}, size={mb} MB')
+                total_size += mask.nbytes
+        return total_size
+
+    def listener(self, ds, path, modality):
+        logger.info('inside of the bold-proc topic')
+        key = int(ds.InstanceNumber)
+        is_multi_echo, is_TE2 = self.check_echo(ds)
+        if is_multi_echo is True and is_TE2 is False:
+            os.remove(path)
+            return
+        if is_multi_echo:
+            key = self.get_new_key(key)
+            logger.info(f'new multi-echo key: {key}')
+        self._instances[key] = {
+            'path': path,
+            'volreg': None,
+            'nii_path': None
+        }
+        self._slice_means[key] = {
+            'path': path,
+            'slice_means': None,
+            'mask_threshold': None,
+            'mask': None
+        }
+        logger.debug('current state of instances')
+        logger.debug(json.dumps(self._instances, default=list, indent=2))
+
+        tasks = self.check_volreg(key)
+        logger.debug('publishing message to volreg topic with the following tasks')
+        logger.debug(json.dumps(tasks, indent=2))
+        pub.sendMessage('volreg', tasks=tasks)
+        logger.debug(f'publishing message to params topic')
+        pub.sendMessage('params', ds=ds, modality=modality)
+
+        logger.debug(f'after volreg')
+        logger.debug(json.dumps(self._instances, indent=2))
+        project = ds.get('StudyDescription', '[STUDY]')
+        session = ds.get('PatientID', '[PATIENT]')
+        scandesc = ds.get('SeriesDescription', '[SERIES]')
+        scannum = ds.get('SeriesNumber', '[NUMBER]')
+        subtitle_string = f'{project} • {session} • {scandesc} • {scannum}'
+        num_vols = ds[(0x0020, 0x0105)].value
+        if self._debug_display:
+            pub.sendMessage('plot', instances=self._instances, subtitle_string=subtitle_string)
+        elif num_vols == key:
+            pub.sendMessage('plot', instances=self._instances, subtitle_string=subtitle_string)
+
[…]
