### 2025-05-16 9c5ee83
get modality types from config file, only start those processors
