### 2025-05-16 51743ba
config checks table position for both localizer and bold
