### 2025-05-19 d21a230
sanity checking config file, remove unused imports, clear up error messages
