+++ scanbuddy/proc/localizer.py
@@ -1,19 +1,5 @@
-import os
-import sys
-import pdb
-import time
-import math
-import json
-import shutil
-import psutil
 import logging
-import datetime
-import threading
-import numpy as np
 from pubsub import pub
-from pathlib import Path
-from sortedcontainers import SortedDict
-from scanbuddy.proc.snr import SNR
 
 logger = logging.getLogger(__name__)
 
