### 2025-05-20 d150916
moving dicom reset message to start rather than start to ensure scanbuddy objects are clear when a new modality scan comes in
