### 2025-05-20 1d82578
modify instruction message for table position error
