### 2025-05-20 856105c
small edits to config file
