+++ example-config.yaml
@@ -63,6 +63,6 @@ modalities:
                     Detected an issue with the table position.
 
                     1. Slide the table position to home.
-                    2. Re-align to iso-center with laser.
+                    2. Realign to isocenter with laser.
                     3. Delete all scans currently in the queue.
-                    4. Re-run localizer and scout, pull over protocal again.
+                    4. Rerun localizer and scout, pull over entire protocal again.
