### 2025-05-20 e4ef880
dont need --debug-display in command anymore, now in docker compose file
