### 2025-05-23 f91f833
beef up simulator.py
