+++ scanbuddy-docker-compose/docker-compose-devel.yaml
@@ -2,5 +2,6 @@ services:
   scanbuddy-web:
     volumes:
       - ../scripts/start.py:/sw/miniforge/envs/python3.13t/bin/start.py
+      - ../scripts/simulator.py:/sw/miniforge/envs/python3.13t/bin/simulator.py
       - ../scanbuddy:/sw/miniforge/envs/python3.13t/lib/python3.13t/site-packages/scanbuddy
     command: ["-c", "/etc/scanbuddy.yaml", "--folder", "/data", "--host", "0.0.0.0", "--broker-host", "scanbuddy-broker"]
