### 2025-05-27 2ec7bc8
add logging for broker
