### 2025-05-28 #8119 @danielasay
Hi @sappelhoff,

Happy to do that!

@dmd has successfully gotten scanbuddy working at his imaging center. Dan has also contributed to the project (3 commits) so he may not be eligible for peer-review.

@dmd, would you be willing to comment on the installation and functionality of scanbuddy at McLean? Thanks so much!!
