### 2025-06-05 118d9f1
add samba container option that works on macos
