### 2025-06-05 56ee169
found a way to use single samba container for Linux and macOS
