+++ scanbuddy-docker-compose/docker-compose.yaml
@@ -57,10 +57,7 @@ services:
     ports:
       - "445:445"
     environment:
-      SAMBA_CONF_LOG_LEVEL: 3
-      WSDD2_DISABLE: 1
-      AVAHI_DISABLE: 1
-      NETBIOS_DISABLE: 1
+      SAMBA_CONF_LOG_LEVEL: 1
       ACCOUNT_scanbuddy: ${SAMBA_PASS}
       UID_scanbuddy: ${UID}
       GROUP_scanbuddy: ${GID}
@@ -72,36 +69,16 @@ services:
           browsable = yes
           create mode = 0660
           directory mode = 0770
-        [bold]
+          vfs objects = catia fruit
+        [localizer]
           path = /data/localizer
           read only = no
           writable = yes
           browsable = yes
           create mode = 0660
           directory mode = 0770
+          vfs objects = catia fruit
     volumes:
       - ${INCOMING_DIR}:/data
     profiles: [samba]
-
-  scanbuddy-samba-macos:
-    build:
-      context: ./samba
-      args:
-        SAMBA_USER: ${SAMBA_USER}
-        SAMBA_PASS: ${SAMBA_PASS}
-        UID: ${UID}
-        GID: ${GID}
-    restart: unless-stopped
-    ports:
-      - "445:445"
-    volumes:
-      - ./samba/logs:/var/log/samba
-      - ./samba/smb.conf:/etc/samba/smb.conf
-      - ${INCOMING_DIR}:/data
-    healthcheck:
-      test: smbclient -L '\\localhost\' -U '%' -m SMB3
-      interval: 30s
-      timeout: 10s
-      retries: 3
-    profiles: [samba-macos]
 
