### 2025-06-26 #8119 @danielasay
@sappelhoff Thank you for the nudge! Apologies for the delay. I will check in with the other group that's using Scanbuddy and see if they would be willing to provide feedback.

If they are unable to, is there another plan we could pursue? Thanks!
