### 2025-06-27 #8119 @sappelhoff
Thank you @danielasay. Another way I can think of would be two short (as in concise) videos that walk through / show the installation and some functionality respectively. These would not be intended as "proof", but rather might serve as an additional tutorial material for making this software more approachable for future users. Please let me know about your thoughts on this idea.
