### 2025-07-23 616e644
change debug display to true in example config file
