+++ paper/citations.bib
@@ -41,6 +41,6 @@
 	doi = {10.3389/conf.fninf.2014.08.00049},
 	urldate = {2025-02-18},
 	journal = {Frontiers in Neuroinformatics},
-	author = {Gabriele, Fariello and Victor, Petrov and Timothy, O'Keefe and Garth, Coombs},
+	author = {Gabriele, Fariello and Victor, Petrov and Timothy, O'Keefe and Garth, Coombs and Randy, Buckner},
 	year = {2014},
 }
