+++ docs/source/user_docs.rst
@@ -12,18 +12,32 @@ Let's start with what Scanbuddy is all about. fMRI is a powerful research and cl
 
 To combat subject motion and optimize data quality, motion-correcting software algorithms can be employed in the post-processing stage, as well as data deletion and imputation methods. However, there are instances of subject motion being severe enough to make the dataset unusable. This is where Scanbuddy comes in! Scanbuddy produces motion plots to be viewed by researchers at the time of data acquisition, appearing on screen at the conclusion of fMRI scans. Individual researchers will determine acceptable motion standards. Seeing motion plots at acquisition can help researchers decide if a scan should be re-acquired. You no longer have to wait until data processing to get an idea of how much your subject has or has not moved.
 
-Scanbuddy also provides an estimate of the Signal-to-Noise Ratio (SNR) with the motion plots to give researchers an idea of overall data quality. Scanbuddy does not save motion plots by default and does not store data on its host machine. Scanbuddy will create a new motion plot and compute a new SNR metric for every fMRI scan acquired. Scanbuddy also supports multi-echo BOLD scans with the assumption that the second echo time (TE2) is the TE of interest. Scanbuddy is containerized with Docker and is available on Github Container Repository.
+Scanbuddy also provides an estimate of the Signal-to-Noise Ratio (SNR) with the motion plots to give researchers an idea of overall data quality. Scanbuddy does not save motion plots by default and does not store data on its host machine. Scanbuddy will create a new motion plot and compute a new SNR metric for every fMRI scan acquired. Scanbuddy also supports multi-echo BOLD scans with the assumption that the second echo time (TE2) is the TE of interest. Scanbuddy also checks for common errors involving head coils and table position. For example, if the headcoil is not plugged in completely Scanbuddy will notify the users with a red WARNING screen to notify them. Similarly, if the table position was configured incorrectly, Scanbuddy will display a red WARNING screen notifying the user. Scanbuddy is containerized with Docker and is available on Github Container Registry.
 
 
 What You Will Need
 ^^^^^^^^^^^^^^^^^^
-Scanbuddy should be run on a standalone machine (separate from the scanner host PC) that runs Linux and you have sudo privileges. We've used several distributions of Linux in development (ubuntu, debian, asahi) and all have run Scanbuddy successfully. The Scanbuddy machine should have 16 GB of RAM if possible; it may still work with less memory depending on the size of the data being acquired. The machine should be capable of running a web browser and Docker. You will also need a monitor to connect to the machine to display the motion plots.
+Scanbuddy should be run on a standalone machine (separate from the scanner host PC) that runs Linux and you have sudo privileges. We've used several distributions of Linux in development (ubuntu, debian, asahi) and all have run Scanbuddy successfully. The Scanbuddy machine should have 16 GB of RAM if possible; it may still work with less memory depending on the size of the data being acquired. The machine should be capable of running a web browser and Docker. You will also need a monitor to connect to the machine to display the motion plots. As an additional safety measure, the computer that you're running Scanbuddy on should be plugged into a surge protector.
 
 .. note::
-     Take a look at installing Docker on Linux `here <https://docs.docker.com/engine/install/>`_.
+     
+    | Docker Linux installation: `Docker Engine <https://docs.docker.com/engine/install/>`_
+    | Docker MacOS installation: `Docker Desktop <https://docs.docker.com/desktop/setup/install/mac-install/>`_
+
+
+Two Installation Options
[…]
