### 2025-07-28 e149c1a
add link to installation video for docker compose
