### 2025-07-28 #8119 @danielasay
Hi @sappelhoff, We've finished making a video that details the installation process of Scanbuddy along with a demonstration of its functionality in production https://youtu.be/JbkHnK4dlpo. We've updated the Scanbuddy documentation on readthedocs that links to this video as an installation path.

I also wanted to let you know that I fixed a citation error in the paper/citations.bib file in the Scanbuddy repository. Please let me know if you have any further questions. Thank you for your patience with this review process.
