### 2025-07-28 #8119 @sappelhoff
@danielasay please work on your points from https://github.com/openjournals/joss-reviews/issues/8119#issuecomment-3128653732 and let me know once you are done.

I will then continue with my list.
