### 2025-07-29 #8119 @sappelhoff
> I released a new version of Scanbuddy: 0.3.0

Thanks @danielasay! However, I don't see it on GitHub: https://github.com/harvard-nrg/scanbuddy/tags

Please also make sure that the tag, release, and filed version under zenodo correspond to each other (including the potential `v` prefix)
