### 2025-07-29 #8119 @sappelhoff
@danielasay the archive contains only the README of the repo:

It should instead the entire default branch (often `main`) of the repository instead 🙂
