### 2025-07-29 #8119 @danielasay
@sappelhoff That would require me to make a new version (0.3.1) of the release/Scanbuddy. Would you like me to do that? We just had the README in this JOSS publication, but I understand if you'd like me to do a new release. https://zenodo.org/records/14605672
