### 2025-07-29 #8119 @danielasay
@sappelhoff Sounds good! Just took care of that.

New doi: https://doi.org/10.5281/zenodo.16579452
New scanbuddy tag on github: https://github.com/harvard-nrg/scanbuddy/releases/tag/0.3.1
