+++ docs/source/index.rst
@@ -6,7 +6,7 @@
 Welcome to Scanbuddy's documentation!
 =====================================
 
-Scanbuddy is an fMRI motion plotting software that allows users to see motion estimates at the time of scan acquisition. Scanbuddy also computes an SNR estimate for the scan. Scanbuddy is written in python and has been containerized to work across platforms, though it was developed for Linux. Enjoy!
+Scanbuddy is an fMRI motion plotting software that allows users to see motion estimates at the time of scan acquisition. Scanbuddy also computes an SNR estimate for the scan. Scanbuddy is written in python and has been containerized to work across platforms, though it was developed for Linux and macOS. Enjoy!
 
 .. toctree::
    :maxdepth: 2
