+++ docs/source/technical_appendix.rst
@@ -19,7 +19,7 @@ Some of you may be wondering why we chose to calculate volume by volume motion r
 
 SNR Calculation
 ^^^^^^^^^^^^^^^
-Scanbuddy's SNR calculation is slice based. It uses coil element information available in the dicom header to determine what the expected voxel intensity should be. Any voxels that are below that threshold get masked out in through every slice in each volume. Then we take the average of each slice from each volume to create a 2D array. We calculate an SNR metric for each slice by dividing the slice mean by the slice standard deviation. Next, we calculate a weighted SNR mean by dividing the slice SNR by the number of voxels in that slice. Finally, we aggregate the weighted slice means across all the volumes and divide it by the total number of voxels.
+Scanbuddy's SNR calculation is slice based. It uses coil element information available in the dicom header to determine what the expected voxel intensity should be. Any voxels that are below that threshold get masked out across every slice in each volume. Then we take the average of each slice from each volume to create a 2D array. We calculate an SNR metric for each slice by dividing the slice mean by the slice standard deviation. Next, we calculate a weighted SNR mean by dividing the slice SNR by the number of voxels in that slice. Finally, we aggregate the weighted slice means across all the volumes and divide it by the total number of voxels.
 
 And voilà! We have an SNR estimate. Scanbuddy starts calculating SNR volume around volume 60, though it may differ depending on the headcoil being used. Scanbuddy will calculate SNR every four volumes after it's initial calculation. We optimized the calculation by having Scanbuddy only recalculate weighted slice means in slices where the estimated mask has changed.
 
