### 2025-08-04 #8119 @danielasay
Hi @sappelhoff,

Thanks again for all your help! Is there a general timeline for when the paper might be published? Thank you!
