### 2025-08-07 f400d22
fix typo in paper.md
