### 2025-08-07 #8119 @danielasay
Hi @rkurchin,

Thank you for your help getting us over the finish line! See the [commit](https://github.com/harvard-nrg/scanbuddy/commit/f400d2296dadae8370e8b8e59f47aa146daa5c71) that fixed the typo you pointed out.

I will defer to the reviewers and @sappelhoff regarding the software installation as part of the review. Please let me know if you need anything else from me.

Thank you!
