### 2025-08-07 #8119 @sappelhoff
> However, I am wondering what stopped reviewers from verifying that the software could be installed?

As far as I can see it, a full installation (and verifying it) would involve setting up at least a rudimentary connection to the scanner. And while some parts (the samba "connection") could also be "mocked", other parts seem more difficult/impossible to verify without a scanner: https://scanbuddy.readthedocs.io/en/main/user_docs.html#configuring-the-plugin

I personally am happy with what @danielasay has provided in terms of documentation / proof of ease of installation, and I was under the impression that the reviewers were happy, too.

Please let us know what you and the other EiCs think @rkurchin, and thank you for taking this on.
