### 2025-08-08 #8119 @danielasay
@rkurchin,

Sounds good! There is a faculty member at BYU with access to the right equipment who has agreed to verify the functional claims. He will be getting Scanbuddy set up sometime over the next several days and will report his findings here when he’s done.
