### 2025-09-15 #8119 @spallenBYU
Thanks, all. I plan to finish my review today. Sept 16 at the latest.

Thanks,
Steven

From: Stefan Appelhoff ***@***.***>
Date: Monday, September 15, 2025 at 3:34 AM
Cc: Steven Allen ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: Scanbuddy: fMRI motion plotting and SNR estimation at scan acquisition (Issue #8119)
[https://avatars.githubusercontent.com/u/9084751?s=20&v=4]sappelhoff left a comment (openjournals/joss-reviews#8119)<https://github.com/openjournals/joss-reviews/issues/8119#issuecomment-3291257312>

@spallenBYU<https://github.com/spallenBYU> do you have an expected timeline for us on performing your review?

@danielasay<https://github.com/danielasay> if you have email contact with Steve Allen, perhaps sending a reminder their way can help. GitHub notifications do sometimes go unnoticed.
