### 2025-09-18 b3c8dff
beta support for vnav images
