### 2025-09-24 #8119 @spallenBYU
Ok. @danielasay , what I recommend is to modify the installation instructions on the readthedocs site. Specifically, maybe between the sections where one can install using Docker Compose and installing services individually, you could put some indication that after instructions in the video are followed one still must configure and test the plugin on the Siemens scanner.
