### 2025-09-26 #8119 @rkurchin
@danielasay just to check, were the only changes as a result of the third review to the documentation? If so, we can proceed without another release tag, etc., but I wanted to explicitly confirm.
